requirements.txt
aiogram==2.20
aiohttp==3.8.1
aiosignal==1.3.1
aiosqlite==0.18.0
async-timeout==4.0.2
attrs==22.2.0
Babel==2.9.1
certifi==2022.12.7
chardet==3.0.4
charset-normalizer==2.1.1
databases==0.6.2
frozenlist==1.3.3
greenlet==2.0.2
idna==3.4
loguru==0.6.0
magic-filter==1.0.9
multidict==6.0.4
ormar==0.12.1
pydantic==1.10.4
python-dotenv==1.0.0
pytz==2023.3
requests==2.28.2
SQLAlchemy==1.4.41
typing_extensions==4.5.0
urllib3==1.26.15
yarl==1.8.2
